[RLS] improve config test and fix bug
